requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
torch==1.13.1
torchvision==0.14.1
scikit-learn==1.3.0
numpy==1.23.5
opencv-python-headless==4.7.0.72
Pillow==9.5.0
gunicorn==20.1.0
python-dotenv==1.0.0
requests==2.31.0
